+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Check for valid data types in service fields
+
 ## [0.0.7] - 2019-04-24
 
 ### Added